+++ lib/groupdate/relation_builder.rb
@@ -203,10 +203,21 @@ module Groupdate
       end
     end
 
+    # basic version of Active Record disallow_raw_sql!
+    # symbol = column (safe), Arel node = SQL (safe), other = untrusted
+    def validate_column(column)
+      # matches table.column and column
+      unless column.is_a?(Symbol) || column.is_a?(Arel::Nodes::SqlLiteral) || /\A\w+(\.\w+)?\z/i.match(column.to_s)
+        warn "[groupdate] Non-attribute argument: #{column}. Use Arel.sql() for known-safe values. This will raise an error in Groupdate 6"
+      end
+    end
+
     # resolves eagerly
     # need to convert both where_clause (easy)
     # and group_clause (not easy) if want to avoid this
     def resolve_column(relation, column)
+      validate_column(column)
+
       node = relation.send(:relation).send(:arel_columns, [column]).first
       node = Arel::Nodes::SqlLiteral.new(node) if node.is_a?(String)
       relation.connection.visitor.accept(node, Arel::Collectors::SQLString.new).value